# [REVIEW]  Use edge_ids directly in uniform sampling call to prevent cost of edge_id lookup